+++ utils.go
@@ -10,6 +10,15 @@ func HandleRecover(handler Handler) {
 	}
 }
 
+// Panic panics if the passed error is not nil.
+//
+// This function is useful with HandleRecover when panic is used as a flow control tool to stop the application.
+func Panic(err error) {
+	if err != nil {
+		panic(err)
+	}
+}
+
 // HandleIfErr handles an error whenever it occurs.
 func HandleIfErr(handler Handler, err error) {
 	if err != nil {